version 0.2.pre

+++ run_3_predict.py
@@ -1,13 +1,7 @@
 # -*- coding:utf-8 -*-
-import os
-import numpy as np
-from config import resultDir, estimator_name, extractor_name,bShuffle
+from model import destdir,clf
 from extracted import extracted
-import pickle
-
-destdir = './_results/%s-%s/'%(extractor_name,estimator_name)
-with open(os.path.join(destdir,'model.pkl'),'rb') as f:
-    clf = pickle.load(f)
+import os
 
 curdir = os.getcwd()
 os.chdir(destdir)
@@ -17,6 +11,8 @@ X_test = extracted['teX']
 testY = clf.predict(X_test)
 fname = 'submission.csv'
 fnamezip = fname+'.zip'
+if os.path.exists(fname):
+    raise RuntimeError("%s exsits, cannot overwrite"%fname)
 with open(fname,'w') as f:
     f.write('instanceId,prob\n')
     for i,y in enumerate(testY):
